requirements.txt
asgiref==3.5.2
asn1crypto==1.4.0
astroid==2.11.5
autopep8==1.5.4
beautifulsoup4==4.11.1
Brlapi==0.8.3
catfish==4.16.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
charset-normalizer==2.0.6
chrome-gnome-shell==0.0.0
click==8.1.2
colorama==0.4.4
cryptography==3.3.1
cupshelpers==1.0
dbus-python==1.2.18
debtags==2.1
defusedxml==0.6.0
dill==0.3.4
distlib==0.3.4
distro==1.7.0
distro-info==1.1
Django==4.0.6
django-allauth==0.44.0
django-debug-toolbar==3.5.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
djlint==0.7.6
filelock==3.7.1
gpg===1.16.0-unknown
gyp==0.1
html5lib==1.1
idna==2.10
importlib-metadata==4.11.3
isort==5.10.1
kali-tweaks==2022.2.0
kazam==1.4.5
lazy-object-proxy==1.7.1
lightdm-gtk-greeter-settings==1.2.2
louis==3.21.0
lxml==4.8.0
mccabe==0.7.0
mechanize==0.4.7
meson==0.61.2
netifaces==0.11.0
oauthlib==3.1.0
packaging==21.3
pathspec==0.9.0
pexpect==4.8.0
Pillow==9.1.1
platformdirs==2.5.2
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.6.0
pycparser==2.20
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.44.1
Pygments==2.11.2
PyGObject==3.42.0
PyJWT==2.0.0
pylint==2.13.9
PyMySQL==1.0.2
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pysmbc==1.0.23
PySocks==1.7.1
python-apt==2.3.0+b1
python-debian==0.1.43
python3-openid==3.2.0
pytz==2020.5
pyxattr==0.7.2
pyxdg==0.27
PyYAML==6.0
ranger-fm==1.9.3
regex==2022.3.15
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.3.1
speedtest-cli==2.1.3
sqlparse==0.4.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.9.2
torbrowser-launcher==0.3.5
tqdm==4.64.0
unattended-upgrades==0.1
urllib3==1.26.2
virtualenv==20.14.1
webencodings==0.5.1
wrapt==1.14.1
xdg==5
youtube-dl==2021.12.17
zipp==3.8.0
